+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/ext/GuiceyExtensionsSupport.java
@@ -27,6 +27,7 @@ import ru.vyarus.dropwizard.guice.test.jupiter.env.field.TestFieldUtils;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.ExtensionConfig;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.track.GuiceyTestTime;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.track.TestExtensionsTracker;
+import ru.vyarus.dropwizard.guice.test.jupiter.ext.log.RecordedLogsSupport;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.mock.MocksSupport;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.rest.RestStubSupport;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.spy.SpiesSupport;
@@ -424,6 +425,7 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
         final List<TestEnvironmentSetup> res = new ArrayList<>();
         if (useDefaultExtensions) {
             res.addAll(tracker.defaultExtensions(
+                    new RecordedLogsSupport(),
                     new RestStubSupport(),
                     new StubsSupport(),
                     new MocksSupport(),